% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWRsim.R
\name{CWRsim}
\alias{CWRsim}
\title{Community Wide evolutionary Rescue simulation (CWR)}
\usage{
CWRsim(tmax, b1, d1, b2, d2, m12, m21, k1, abun_original, interval)
}
\arguments{
\item{tmax}{Arbitrary units of time the simulation should be run}

\item{b1}{Birthrate of the mutants}

\item{d1}{Deathrate of the mutants}

\item{b2}{Birthrate of the residents}

\item{d2}{Death rate of the residents}

\item{m12}{Mutation rate of mutants to residents}

\item{m21}{Mutation rate of residents to mutants}

\item{k1}{Carrying capacity}

\item{abun_original}{A vector of initial abundances for each species such as those created by generate_spat_abund}

\item{interval}{After how many evenst should the population state be saved to the output matrix}
}
\value{
A matrix denoting the abundances of all the species in the community over time
}
\description{
Community Wide evolutionary Rescue simulation (CWR)
}
\examples{
CWRsim(10,0.6 , 0.1, 0.05, 0.1, 0, 0.0005, 16000,
generate_spat_abund(theta = 200,Ivec = rep(40,1),Jvec = c(16000)), 500)

}
