% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWRsimSPEC.R
\name{CWRsimSPEC}
\alias{CWRsimSPEC}
\title{Community Wide evolutionary Rescue simulation (CWR), with certain precise timepoints saved}
\usage{
CWRsimSPEC(
  tmax,
  b1,
  d1,
  b2,
  d2,
  m12,
  m21,
  k1,
  abun_original,
  interval,
  wantedtimes
)
}
\arguments{
\item{tmax}{Arbitrary units of time the simulation should be run}

\item{b1}{Birthrate of the mutants}

\item{d1}{Deathrate of the mutants}

\item{b2}{Birthrate of the residents}

\item{d2}{Death rate of the residents}

\item{m12}{Mutation rate of mutants to residents}

\item{m21}{Mutation rate of residents to mutants}

\item{k1}{Carrying capacity}

\item{abun_original}{A vector of initial abundances for each species such as those created by generate_spat_abund}

\item{interval}{After how many evenst should the population state be saved to the output matrix}

\item{wantedtimes}{The preciese timepoints that should be saved}
}
\value{
A matrix denoting the abundances of all the species in the community over time, with certain precise timpoints saved
}
\description{
Community Wide evolutionary Rescue simulation (CWR), with certain precise timepoints saved
}
\examples{
CWRsimSPEC(10,0.6 , 0.1, 0.05, 0.1, 0, 0.0005, 16000,
generate_spat_abund(theta = 200,Ivec = rep(40,1),Jvec = c(16000)),
 200,c(15,30,50,75,100))

}
