% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_spat_abund.R
\name{generate_spat_abund}
\alias{generate_spat_abund}
\title{Generate a neutral community}
\usage{
generate_spat_abund(theta, Ivec, Jvec)
}
\arguments{
\item{theta}{an integer that gives a value to Theta}

\item{Ivec}{a vector of I values}

\item{Jvec}{a vector of J values}
}
\value{
a vector of species abundances
}
\description{
Uses sampling function for neutral community from Etienne et al. 2007 (Ecology Letters), used to initialize communities.
}
\examples{
generate_spat_abund(theta = 200,Ivec = rep(40,1),Jvec = c(16000))
}
