% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinulsim.R
\name{multinulsim}
\alias{multinulsim}
\title{Multiple Simple Neutral community (SN) simulations}
\usage{
multinulsim(tmax, b1, d1, k1, abun_original, nsim, filenomen, interval)
}
\arguments{
\item{tmax}{Arbitrary units of time the simulation should be run}

\item{b1}{Birthrate}

\item{d1}{Deathrate}

\item{k1}{Carrying capacity}

\item{abun_original}{A vector of initial abundances for each species such as those created by generate_spat_abund}

\item{nsim}{number of simulations that should be performed (integer)}

\item{filenomen}{the name of the plot that is saved}

\item{interval}{After how many evenst should the population state be saved to the output matrix}
}
\value{
this function plots all the results of the simulations in a .tiff file in the current working directory
}
\description{
Multiple Simple Neutral community (SN) simulations
}
\examples{
multinulsim(10, 0.6, 0.1, 16000,
generate_spat_abund(theta = 200,Ivec = rep(40,1),Jvec = c(16000)),
2,"nulsim.tiff",  200)



}
