% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinulsimSPEC.R
\name{multinulsimSPEC}
\alias{multinulsimSPEC}
\title{Multiple Simple Neutral community (SN) simulations, plotting the RAC at specific timepoints and also saving the output matrix for every simulation}
\usage{
multinulsimSPEC(tmax, b1, d1, k1, abun_original, nsim, interval, wantedtimes)
}
\arguments{
\item{tmax}{Arbitrary units of time the simulation should be run}

\item{b1}{Birthrate}

\item{d1}{Deathrate}

\item{k1}{Carrying capacity}

\item{abun_original}{A vector of initial abundances for each species such as those created by generate_spat_abund}

\item{nsim}{number of simulations that should be performed (integer)}

\item{interval}{After how many evenst should the population state be saved to the output matrix}

\item{wantedtimes}{Specific timepoints for which the RAC is to be plotted}
}
\value{
this function plots the RAC of the simulations for the times specified in Wantedtimes as .tiff files in the current working directory also writes a .csv file outputting the poplation matrix for every simulation performed
}
\description{
Multiple Simple Neutral community (SN) simulations, plotting the RAC at specific timepoints and also saving the output matrix for every simulation
}
\examples{
multinulsimSPEC(10, 0.6, 0.1, 1600,
generate_spat_abund(theta = 200,Ivec = rep(40,1),Jvec = c(16000)),
2, 200, c(15,30,50,75,100))



}
