% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisimbvar.R
\name{multisimbvar}
\alias{multisimbvar}
\title{Simulate multiple neutral communities with a variable birthrate (VBN)}
\usage{
multisimbvar(
  tmax,
  b1,
  d1,
  k1,
  bneck,
  kneckstart,
  kneckend,
  abun_original,
  nsim,
  filenomen,
  interval
)
}
\arguments{
\item{tmax}{Arbitrary units of time the simulation should be run}

\item{b1}{Normal birthrate}

\item{d1}{Deathrate}

\item{k1}{Carrying capacity}

\item{bneck}{Birthrate during the bottleneck}

\item{kneckstart}{Timepoint at which the bottleneck starts}

\item{kneckend}{Timepoint at which the bottleneck ends}

\item{abun_original}{A vector of initial abundances for each species such as those created by generate_spat_abund}

\item{nsim}{number of simulations that should be performed}

\item{filenomen}{the name of the plot that is saved}

\item{interval}{After how many evenst should the population state be saved to the output matrix}
}
\value{
this function plots all the results of the simulations in a .tiff file in the current working directory
}
\description{
Simulate multiple neutral communities with a variable birthrate (VBN)
}
\examples{
multisimbvar( 10, 0.6, 0.1, 16000, 0.05,0, 5,
generate_spat_abund (theta = 200,Ivec = rep(40,1),Jvec = c(16000)),
2, "bvar.tiff", 200)


}
