% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisimbvarSPEC.R
\name{multisimbvarSPEC}
\alias{multisimbvarSPEC}
\title{Simulate multiple neutral communities with a variable birthrate (VBN) saving specific timepoints and also saving the output matrix for every simulation}
\usage{
multisimbvarSPEC(
  tmax,
  b1,
  d1,
  k1,
  bneck,
  kneckstart,
  kneckend,
  abun_original,
  nsim,
  interval,
  wantedtimes
)
}
\arguments{
\item{tmax}{Arbitrary units of time the simulation should be run}

\item{b1}{Normal birthrate}

\item{d1}{Deathrate}

\item{k1}{Carrying capacity}

\item{bneck}{Birthrate during the bottleneck}

\item{kneckstart}{Timepoint at which the bottleneck starts}

\item{kneckend}{Timepoint at which the bottleneck ends}

\item{abun_original}{A vector of initial abundances for each species such as those created by generate_spat_abund}

\item{nsim}{number of simulations that should be performed}

\item{interval}{After how many evenst should the population state be saved to the output matrix}

\item{wantedtimes}{Specific timepoints for which the RAC is to be plotted}
}
\value{
this function plots the RAC of the simulations for the times specified in Wantedtimes as .tiff files in the current working directory also writes a .csv file outputting the poplation matrix for every simulation performed
}
\description{
Simulate multiple neutral communities with a variable birthrate (VBN) saving specific timepoints and also saving the output matrix for every simulation
}
\examples{
multisimbvarSPEC(10, 0.6, 0.1, 16000, 0.05, 0, 5,
generate_spat_abund (theta = 200,Ivec = rep(40,1),Jvec = c(16000)),2,
200, c(15,30,50,75,100))


}
\author{
Timo van Eldijk
}
