% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nulsimSPEC.R
\name{nulsimSPEC}
\alias{nulsimSPEC}
\title{Simple Neutral community (SN) saving certain precise timepoints}
\usage{
nulsimSPEC(tmax, b1, d1, k1, abun_original, interval, wantedtimes)
}
\arguments{
\item{tmax}{Arbitrary units of time the simulation should be run}

\item{b1}{Birthrate}

\item{d1}{Deathrate}

\item{k1}{Carrying capacity}

\item{abun_original}{A vector of initial abundances for each species such as those created by generate_spat_abund}

\item{interval}{After how many evenst should the population state be saved to the output matrix}

\item{wantedtimes}{The specific timepoints that should be included in the final matrix}
}
\value{
A matrix denoting the abundances of all the species in the community over time, with certain precise timepoints saved (wantedtimes)
}
\description{
Simple Neutral community (SN) saving certain precise timepoints
}
\examples{
nulsimSPEC(10, 0.6, 0.1, 16000, generate_spat_abund(theta = 200,Ivec = rep(40,1),
Jvec = c(16000)), 20000000, c(1,3,5))

}
\author{
Timo van Eldijk
}
