% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simbvar.R
\name{simbvar}
\alias{simbvar}
\title{Simulate a neutral community with a variable birthrate (VBN)}
\usage{
simbvar(tmax, b1, d1, k1, bneck, kneckstart, kneckend, abun_original, interval)
}
\arguments{
\item{tmax}{Arbitrary units of time the simulation should be run}

\item{b1}{Normal birthrate}

\item{d1}{Deathrate}

\item{k1}{Carrying capacity}

\item{bneck}{Birthrate during the bottleneck}

\item{kneckstart}{Timepoint at which the bottleneck starts}

\item{kneckend}{Timepoint at which the bottleneck ends}

\item{abun_original}{A vector of initial abundances for each species such as those created by generate_spat_abund}

\item{interval}{After how many evenst should the population state be saved to the output matrix}
}
\value{
A matrix denoting the abundances of all the species in the community over time
}
\description{
Simulate a neutral community with a variable birthrate (VBN)
}
\examples{
simbvar( 10, 0.6, 0.1, 16000, 0.05,0, 5,
generate_spat_abund(theta = 200,Ivec = rep(40,1),Jvec = c(16000)), 200)

}
